+++ src/Content/Backend/NV.Templates.Backend.Web/Framework/OpenApi/OpenApiServiceCollectionExtensions.cs
@@ -1,4 +1,6 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
+using Microsoft.AspNetCore.Mvc.ApiExplorer;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Options;
 using NSwag;
@@ -18,35 +20,43 @@ namespace Microsoft.Extensions.DependencyInjection
 #if Auth
             services.BindOptionsToConfigurationAndValidate<OpenApiSecurityScheme>(configuration);
 #endif
-            services.AddOpenApiDocument((document, sp) =>
+            var initProvider = services.BuildServiceProvider();
+            var serviceInfo = initProvider.GetRequiredService<IApplicationInfo>();
+            var apiVersionDescriptionProvider = initProvider.GetRequiredService<IApiVersionDescriptionProvider>();
+
+            foreach (var apiVersionDescription in apiVersionDescriptionProvider.ApiVersionDescriptions.OrderByDescending(x => x.GroupName))
             {
-                var appInfo = sp.GetRequiredService<IApplicationInfo>();
-
-                document.DocumentName = "v1";
-                document.Title = $"{appInfo.Name} ({appInfo.Environment})";
-                document.Version = appInfo.Version;
-                document.Description = $"<i>For 3rd party licenses see <a href='{AttributionsHandler.Path}'>attributions</a>.</i>";
-                document.ApiGroupNames = new[] { "v1" };
-
-                // TODO: This will need updating when NSwag support System.Text.Json properly.
-                document.SerializerSettings = new Newtonsoft.Json.JsonSerializerSettings
+                services.AddOpenApiDocument((document, sp) =>
                 {
-                    ContractResolver = new Newtonsoft.Json.Serialization.DefaultContractResolver
+                    var appInfo = sp.GetRequiredService<IApplicationInfo>();
+
+                    document.DocumentName = apiVersionDescription.GroupName;
+                    document.Title = $"{serviceInfo.Name} ({serviceInfo.Environment})";
+                    document.Version = serviceInfo.Version;
+                    document.Description = $"<i>For 3rd party licenses see <a href='{AttributionsHandler.Path}'>attributions</a>.</i>";
+                    document.ApiGroupNames = new[] { apiVersionDescription.GroupName };
+
+                    // TODO: This will need updating when NSwag support System.Text.Json properly.
+                    document.SerializerSettings = new Newtonsoft.Json.JsonSerializerSettings
                     {
-                        NamingStrategy = new Newtonsoft.Json.Serialization.CamelCaseNamingStrategy(),
-                    },
-                    Converters = new List<Newtonsoft.Json.JsonConverter> { new Newtonsoft.Json.Converters.StringEnumConverter() },
-                };
+                        ContractResolver = new Newtonsoft.Json.Serialization.DefaultContractResolver
+                        {
+                            NamingStrategy = new Newtonsoft.Json.Serialization.CamelCaseNamingStrategy(),
+                        },
+                        Converters = new List<Newtonsoft.Json.JsonConverter> { new Newtonsoft.Json.Converters.StringEnumConverter() },
+                    };
 
-                document.SchemaNameGenerator = new CustomSchemaNameGenerator();
+                    document.SchemaNameGenerator = new CustomSchemaNameGenerator();
 
-                document.DocumentProcessors.Add(new HealthChecksDocumentProcessor());
-                document.OperationProcessors.Add(new CommonHeadersOperationProcessor());
+                    document.DocumentProcessors.Add(new HealthChecksDocumentProcessor());
+                    document.OperationProcessors.Add(new CommonHeadersOperationProcessor());
+
 #if Auth
-                var openApiSecurityScheme = sp.GetRequiredService<IOptions<OpenApiSecurityScheme>>().Value;
-                document.AddSecurity("OAuth2", openApiSecurityScheme);
+                    var openApiSecurityScheme = sp.GetRequiredService<IOptions<OpenApiSecurityScheme>>().Value;
+                    document.AddSecurity("OAuth2", openApiSecurityScheme);
 #endif
-            });
+                });
+            }
 
             return services;
         }